refactor: convert serial name to serialized name
